BUG/ProductService - 주문 생성시 상품조회 에러 해결, 카프카 메세지 구독 에러 해결, 테스트시 레디스 테스트 제외

+++ product-service/src/main/java/com/palja/product_service/application/dto/res/ProductInfoForOrderRes.java
@@ -14,8 +14,8 @@ public class ProductInfoForOrderRes {
     private UUID productId;
     private UUID companyUserId;
     private String productName;
-    private BigDecimal price;
-    private int stockQuantity;
+    private Long price;
+    private Long stockQuantity;
 
     public static ProductInfoForOrderRes fromDto(ProductInfoForOrderDto dto) {
